Bugfix for failing C compset DEBUG tests
When water isotopes were added to the coupler, the runoff -> ocean fields
changed -- this is fine in coupled and ocean / ice cases, but the new isotope
fields should not be used in ocean-only compsets. To get around this, I
introduced a new module variable (lwiso_runoff) in prep_ocn_mod.F90 that is
true if index_x2o_Foxx_rofl_16O is non-zero; if lwiso_runoff is true then
seq_map_map is called with a full list of fields (i.e. all the water isotopes)
otherwise it is just called with Forr_rofl and Forr_rofi.